Fixed grid column issue for IE7 & 8
After bringing in the original responsive-gs files, respond.js was
removed for simplicity (and due to clients generally not needing
responsive elements for older browsers). However, the CSS was then not
updated to force column sizes for IE7 and 8. This commit moves the
column sizing CSS out of the media query, and fixes clearfixes in IE7.